requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
attrs==24.2.0
av==12.3.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.3
dnspython==2.7.0
et_xmlfile==2.0.0
etelemetry==0.3.1
extra-streamlit-components==0.1.71
fastexcel==0.12.0
filelock==3.16.1
fitz==0.0.1.dev2
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.12.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
nltk==3.9.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
plotly==5.24.1
polars==1.12.0
protobuf==5.28.3
prov==2.0.1
puremagic==1.28
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
pydot==3.0.2
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pylibsrtp==0.10.0
pymongo==4.10.1
PyMuPDF==1.24.13
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzbar==0.1.9
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resend==2.4.0
rich==13.9.3
rpds-py==0.20.0
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
streamlit-authenticator==0.4.1
streamlit-camera-input-live==0.2.0
streamlit-drawable-canvas==0.9.3
streamlit-lottie==0.0.5
streamlit-webrtc==0.47.9
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
watchdog==5.0.3
XlsxWriter==3.2.0
